More validation for imagepreviewer

+++ OSIRT/UI/ImagePreviewerForm.cs
@@ -5,11 +5,13 @@ using OSIRT.Helpers;
 using OSIRT.Loggers;
 using OSIRT.Properties;
 using System;
+using System.Collections.Generic;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Drawing;
 using System.IO;
 using System.Linq;
+using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
@@ -25,6 +27,7 @@ namespace OSIRT.UI
         private readonly int MaxImageHeight = 12500;
         private BackgroundWorker hashCalcBackgroundWorker;
         private CannotOpenImagePanel cantOpenPanel;
+        private ToolTip tooltip;
 
         public string FileName { get { return uiImageNameComboBox.Text; } }
         public string FileExtension { get { return uiFileExtensionComboBox.Text; } }
@@ -33,8 +36,10 @@ namespace OSIRT.UI
         public ImagePreviewerForm()
         {
             InitializeComponent();
-
+            
         }
+
+
 
         private void BackgroundWorker_RunWorkerCompleted(object sender, RunWorkerCompletedEventArgs e)
         {
@@ -60,7 +65,11 @@ namespace OSIRT.UI
         {
             this.imagePath = Constants.TempImgFile;
             this.details = details;
+            tooltip = new ToolTip();
+            uiNoteSpellBox.KeyUp += UiNoteSpellBox_KeyUp;
         }
+
+
 
         private void InitialiseBackgroundWorker()
         {
@@ -144,32 +153,20 @@ namespace OSIRT.UI
             uiFileExtensionComboBox.Items.Add(SaveableFileTypes.Pdf);
             uiFileExtensionComboBox.SelectedIndex = 0;
 
-            //TODO: Remove extension from file name, too.
-            string path = Path.Combine(Constants.ContainerLocation, Constants.Directories.GetSpecifiedCaseDirectory(CaseDirectory.Screenshots));
-            string[] files =  Directory.GetFiles(path).Select(p => Path.GetFileName(p)).ToArray();
-            uiImageNameComboBox.Items.AddRange(files);
+            PopulateComboboxWithFiles();
 
         }
 
+        private void PopulateComboboxWithFiles()
+        {
+            string path = Path.Combine(Constants.ContainerLocation, Constants.Directories.GetSpecifiedCaseDirectory(CaseDirectory.Screenshots));
+            string[] files = Directory.GetFiles(path).Select(p => Path.GetFileNameWithoutExtension(p)).ToArray();
+            uiImageNameComboBox.Items.AddRange(files);
+        }
+
+
         private void uiOKButton_Click(object sender, EventArgs e)
         {
-
-
-            if (!IsValidFileName())
-            {
-                //display lable instead. MessageBox for now
-                MessageBox.Show("Please enter a valid file name. This one may already exist");
-                return;
-            }
-
-
-            if (string.IsNullOrWhiteSpace(Note))
-            {
-                //display lable instead. MessageBox for now
-                MessageBox.Show("Must enter a note.");
-                return;
-            }
-
 
             if (FileExtension == SaveableFileTypes.Pdf)
             {
@@ -186,11 +183,11 @@ namespace OSIRT.UI
 
 
             Log();
-
-            //TODO: delete cache
-            //TODO: trigger dialogresult OK
+            DialogResult = DialogResult.OK;
+            Close();
 
         }
+
 
         private void Log()
         {
@@ -211,7 +208,6 @@ namespace OSIRT.UI
                 using (MagickImage image = new MagickImage(imagePath))
                 {
                     image.Format = MagickFormat.Pdf;
-                    //TODO: may not always be "screenshots" for specified case directory (may be snippet, for example). Enum?
                     pathToSave = Path.Combine(Constants.ContainerLocation, Constants.Directories.GetSpecifiedCaseDirectory(CaseDirectory.Screenshots), fileName + SaveableFileTypes.Pdf);
                     image.Write(pathToSave);
                 }
@@ -228,10 +224,13 @@ namespace OSIRT.UI
             {
                 //delete temp pdf file
                 if (thrown)
-                    File.Delete(pathToSave);
+                {
+                    if (File.Exists(pathToSave))
+                    {
+                        File.Delete(pathToSave);
+                    }
+                }
             }
-
-
         }
 
         private void SaveAsPNG(string name)
@@ -256,24 +255,55 @@ namespace OSIRT.UI
 
         private bool IsValidFileName()
         {
-            //TODO: needs to fire when the combobox has changed index
-            string path = Path.Combine(Constants.ContainerLocation, Constants.Directories.GetSpecifiedCaseDirectory(CaseDirectory.Screenshots), FileName + FileExtension);
-            return !(File.Exists(path) || string.IsNullOrWhiteSpace(FileName));
+            bool valid = false;
+            //must check this first, as trying to use Path.Combine with an illegal file char will thrown an argument exception
+            if(!string.IsNullOrWhiteSpace(FileName) && OsirtHelper.IsValidFilename(FileName))
+            {
+                string path = Path.Combine(Constants.ContainerLocation, Constants.Directories.GetSpecifiedCaseDirectory(CaseDirectory.Screenshots), FileName + FileExtension);
+                if (!File.Exists(path))
+                {
+                    valid = true;
+                }
+            }
+            return valid;
         }
 
 
-
+        //TODO: better name for this method
         private void CheckValidFileName()
         {
-
+            string tootipText = "";
             if (IsValidFileName())
             {
                 uiDoesFileExistPictureBox.Image = Properties.Resources.ok;
+                tootipText = "Filename OK";
             }
             else
             {
-                uiDoesFileExistPictureBox.Image = Properties.Resources.file_clash;
+                uiDoesFileExistPictureBox.Image = Properties.Resources.invalid_entry;
+                tootipText = "Filename is not valid. File with that name may already exists, or filename contains illegal characters.";
             }
+
+            tooltip.SetToolTip(uiDoesFileExistPictureBox, tootipText);
+        }
+
+        private void CheckValidNoteEntry()
+        {
+            if (string.IsNullOrWhiteSpace(Note))
+            {
+                tooltip.SetToolTip(uiNotePictureBox, "You must enter a note.");
+                uiNotePictureBox.Image = Properties.Resources.invalid_entry;
+            }
+            else
+            {
+                tooltip.SetToolTip(uiNotePictureBox, "Note OK");
+                uiNotePictureBox.Image = Properties.Resources.ok;
+            }
+        }
+
+        private void UiNoteSpellBox_KeyUp(object sender, System.Windows.Input.KeyEventArgs e)
+        {
+            CheckValidNoteEntry();
         }
 
         private void uiFileExtensionComboBox_SelectedIndexChanged(object sender, EventArgs e)
